game: add placeholder projects for components
game-events: Since events are sent from client to the server and vice versa, both sides have to know what kind of events there are. This library makes them reusable.
game-logic: The client uses the logic to control the UI and tell the player what moves are allowed and which ones aren't. Since the player could manipulate the client, we have to check each move again on the server, just to make sure.
game-webgl: Due to separation of concerns the actual implementation of the game in WebGL is factored out into a separate library to make it easier to reason about.