server/webapp: fix docker files by including shared local npm packages
The main problem is the docker build context which doesn't allow files/folders to be accessed that are outside this build context.
This poses a problem in our case because the shared libraries are one folder above the build context.
Moving everything root would mean the entire repository would have to be sent as build context to fly.io each time a change is made which is insane.
The solution simply clones the current version from our github repository and uses that as basis.